+++ lib/db_agent/viewpoint.rb
@@ -1,0 +1,2 @@
+require_relative 'viewpoint/base'
+require_relative 'viewpoint/delegate'